Make the vertex shader input vertex an unsigned int. Make define_vertex_spec_index take some buffer_size_t's for some params instead to avoid the possibility of overflow.

+++ assets/shaders/sector.vert
@@ -2,7 +2,7 @@
 
 #include "csm/shadow.vert"
 
-layout(location = 0) in vec3 vertex_pos_world_space;
+layout(location = 0) in uvec3 vertex_pos_world_space;
 layout(location = 1) in uint face_info_bits;
 
 flat out uint face_id;
